Apollo static analysis - Group A
:Release Notes:
CERT INT31-C fix

:Detailed Notes:
CERT INT31-C fix

:Testing Performed:
Local build OK

:QA Notes:
NA

:Issues Addressed:
[WRP-27386] Apollo static analysis - Group A

+++ src/plugins/notifier/pdm/pdm_client.cpp
@@ -184,8 +184,8 @@ bool PDMClient::getDeviceListCallback(const char *message)
     std::vector<DEVICE_LIST_T> devList;
     for (auto jDevice : jPayload["videoDeviceList"])
     {
-        VideoDevice device          = jDevice;
-        unsigned int subdeviceCount = device.subDeviceList.size();
+        VideoDevice device           = jDevice;
+        unsigned long subdeviceCount = device.subDeviceList.size();
         for (auto subdevice : device.subDeviceList)
         {
             DEVICE_LIST_T devInfo;
